// cargodelivery-front/generated/models/FuelStation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Point } from './Point';
import {
    PointFromJSON,
    PointFromJSONTyped,
    PointToJSON,
} from './Point';

/**
 * 
 * @export
 * @interface FuelStation
 */
export interface FuelStation {
    /**
     * 
     * @type {string}
     * @memberof FuelStation
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof FuelStation
     */
    description: string;
    /**
     * 
     * @type {Point}
     * @memberof FuelStation
     */
    coordinates: Point;
    /**
     * 
     * @type {string}
     * @memberof FuelStation
     */
    id: string;
}

/**
 * Check if a given object implements the FuelStation interface.
 */
export function instanceOfFuelStation(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "description" in value;
    isInstance = isInstance && "coordinates" in value;
    isInstance = isInstance && "id" in value;

    return isInstance;
}

export function FuelStationFromJSON(json: any): FuelStation {
    return FuelStationFromJSONTyped(json, false);
}

export function FuelStationFromJSONTyped(json: any, ignoreDiscriminator: boolean): FuelStation {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'],
        'coordinates': PointFromJSON(json['coordinates']),
        'id': json['id'],
    };
}

export function FuelStationToJSON(value?: FuelStation | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'description': value.description,
        'coordinates': PointToJSON(value.coordinates),
        'id': value.id,
    };
}